# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/jobs/{id}": {
            "get": {
                "description": "Retrieves the current status and progress of a processing job. The response includes processing statistics (processed_count, total_count, gemini_calls, flagged_count) and the moderation criteria used for this job.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID from upload response",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job status with progress metrics and criteria",
                        "schema": {
                            "$ref": "#/definitions/model.Job"
                        }
                    },
                    "400": {
                        "description": "Bad request - job id required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tweets": {
            "get": {
                "description": "Retrieves a paginated list of flagged tweets. **IMPORTANT:** After uploading an archive, use the job_id from the upload response to view tweets from that specific job. Example: GET /tweets?job_id=abc123",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tweets"
                ],
                "summary": "List flagged tweets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID from upload response (required to view tweets from your archive)",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of flagged tweets with pagination",
                        "schema": {
                            "$ref": "#/definitions/model.ListTweetsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tweets/export": {
            "get": {
                "description": "Exports URLs of flagged tweets for manual deletion. **IMPORTANT:** Use job_id from upload response to export tweets from your processed archive. Example: GET /tweets/export?job_id=abc123\u0026format=csv",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/csv"
                ],
                "tags": [
                    "tweets"
                ],
                "summary": "Export flagged tweet URLs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID from upload response (required to export tweets from your archive)",
                        "name": "job_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "json",
                            "csv"
                        ],
                        "type": "string",
                        "description": "Export format: json or csv (default json)",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV response with one URL per line (when format=csv)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tweets/upload": {
            "post": {
                "description": "Uploads a Twitter/X archive for processing. Accepts either: (1) ZIP archive file containing data/tweet.js, or (2) unzipped tweet.js/tweets.js file directly. Optional: Provide custom moderation criteria as JSON in the \"criteria\" form field. The criteria will be used by Gemini AI to evaluate tweets. Example criteria format: {\"criteria\": {\"forbidden_words\": [\"crypto\", \"NFT\", \"hustlegrindset\"], \"professional_check\": true, \"tone\": \"respectful and thoughtful\", \"exclude_politics\": true}}. Returns a job_id that you can use to check processing status and view flagged tweets. Example workflow: 1) POST /tweets/upload with file (and optional criteria), 2) Get job_id from response, 3) Check status at GET /jobs/{job_id}, 4) View tweets at GET /tweets?job_id={job_id}",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tweets"
                ],
                "summary": "Upload tweet archive",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Twitter/X archive (ZIP file or unzipped tweet.js/tweets.js file, max 500MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional JSON string with moderation criteria. Example: {\\",
                        "name": "criteria",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Upload successful - save the job_id from this response to query tweets later",
                        "schema": {
                            "$ref": "#/definitions/model.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - file required or too large, or invalid criteria JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tweets/view": {
            "get": {
                "description": "Retrieves details of a specific flagged tweet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tweets"
                ],
                "summary": "Get flagged tweet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flagged tweet ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flagged tweet details",
                        "schema": {
                            "$ref": "#/definitions/model.FlaggedTweet"
                        }
                    },
                    "400": {
                        "description": "Bad request - id parameter required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tweet not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ExportTweetsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.FlaggedTweet": {
            "description": "A tweet that was flagged for deletion based on moderation criteria. Includes the tweet content, reason for flagging, and URL for manual deletion.",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "When the tweet was originally posted",
                    "type": "string",
                    "example": "2023-04-15T10:30:00Z"
                },
                "filter_reason": {
                    "description": "Which filter caught it: \"llm_scorer\" (Gemini) or \"blocked_phrase\" (deterministic)",
                    "type": "string",
                    "enum": [
                        "llm_scorer",
                        "blocked_phrase"
                    ],
                    "example": "llm_scorer"
                },
                "flagged_at": {
                    "description": "When the tweet was flagged by the system",
                    "type": "string",
                    "example": "2024-01-18T12:45:30Z"
                },
                "id": {
                    "description": "Internal unique identifier",
                    "type": "string",
                    "example": "internal-uuid-123"
                },
                "job_id": {
                    "description": "ID of the job that processed this tweet",
                    "type": "string",
                    "example": "job-abc123"
                },
                "reason": {
                    "description": "Explanation of why the tweet was flagged",
                    "type": "string",
                    "example": "Contains severe profanity used aggressively"
                },
                "score": {
                    "description": "Confidence score (0.0-1.0, higher = more likely to flag)",
                    "type": "number",
                    "example": 0.85
                },
                "text": {
                    "description": "Full text of the tweet",
                    "type": "string",
                    "example": "This is a problematic tweet"
                },
                "tweet_id": {
                    "description": "Original Twitter/X tweet ID",
                    "type": "string",
                    "example": "1649447064454017026"
                },
                "url": {
                    "description": "Tweet URL for manual deletion",
                    "type": "string",
                    "example": "https://twitter.com/i/web/status/1649447064454017026"
                }
            }
        },
        "model.Job": {
            "description": "Job status and progress information. Includes processing metrics and the moderation criteria used.",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "When the job was created",
                    "type": "string",
                    "example": "2024-01-18T10:30:00Z"
                },
                "criteria": {
                    "description": "Moderation criteria used for this job (if custom criteria provided)",
                    "$ref": "#/definitions/model.ModerationCriteria"
                },
                "error": {
                    "description": "Error message (only present if status is \"failed\")",
                    "type": "string",
                    "example": "Failed to parse archive"
                },
                "file_id": {
                    "description": "ID of the uploaded archive file",
                    "type": "string",
                    "example": "file-xyz"
                },
                "flagged_count": {
                    "description": "Number of tweets flagged for deletion",
                    "type": "integer",
                    "example": 45
                },
                "gemini_calls": {
                    "description": "Number of Gemini API calls made for this job",
                    "type": "integer",
                    "example": 1200
                },
                "id": {
                    "description": "Unique job identifier",
                    "type": "string",
                    "example": "abc123"
                },
                "processed_count": {
                    "description": "Number of tweets processed so far",
                    "type": "integer",
                    "example": 1500
                },
                "result": {
                    "description": "Final job result (only present when status is \"done\")",
                    "$ref": "#/definitions/model.JobResult"
                },
                "status": {
                    "description": "Current job status",
                    "type": "string",
                    "enum": [
                        "queued",
                        "running",
                        "done",
                        "failed",
                        "paused_quota",
                        "paused_manual"
                    ],
                    "example": "running"
                },
                "total_count": {
                    "description": "Total number of tweets to process",
                    "type": "integer",
                    "example": 2000
                },
                "updated_at": {
                    "description": "When the job was last updated",
                    "type": "string",
                    "example": "2024-01-18T10:45:00Z"
                }
            }
        },
        "model.JobResult": {
            "type": "object",
            "properties": {
                "flagged_count": {
                    "description": "Number of tweets flagged",
                    "type": "integer",
                    "example": 45
                }
            }
        },
        "model.ListTweetsResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "type": "object",
                    "properties": {
                        "page": {
                            "type": "integer"
                        },
                        "page_size": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        },
                        "total_pages": {
                            "type": "integer"
                        }
                    }
                },
                "tweets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FlaggedTweet"
                    }
                }
            }
        },
        "model.ModerationCriteria": {
            "description": "Custom moderation rules passed to Gemini AI to guide tweet evaluation. All fields are optional. If not provided, default criteria will be used.",
            "type": "object",
            "properties": {
                "exclude_harassment": {
                    "description": "If true, flags harassment or targeted abuse (default: true)",
                    "type": "boolean",
                    "example": true
                },
                "exclude_hate_speech": {
                    "description": "If true, flags hate speech or discrimination targeting groups (default: true)",
                    "type": "boolean",
                    "example": true
                },
                "exclude_politics": {
                    "description": "If true, flags political content or statements",
                    "type": "boolean",
                    "example": true
                },
                "exclude_profanity": {
                    "description": "If true, flags severe profanity used aggressively (default: true)",
                    "type": "boolean",
                    "example": true
                },
                "exclude_threats": {
                    "description": "If true, flags threats of violence or harm (default: true)",
                    "type": "boolean",
                    "example": true
                },
                "forbidden_words": {
                    "description": "Words or phrases that should be flagged (case-insensitive)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "crypto",
                        "NFT",
                        "hustlegrindset"
                    ]
                },
                "professional_check": {
                    "description": "If true, flags unprofessional language or tone",
                    "type": "boolean",
                    "example": true
                },
                "tone": {
                    "description": "Expected tone. Tweets deviating significantly from this tone will be flagged",
                    "type": "string",
                    "example": "respectful and thoughtful"
                }
            }
        },
        "model.StoredFile": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "original_name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "storage_backend": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.UploadResponse": {
            "type": "object",
            "properties": {
                "file": {
                    "$ref": "#/definitions/model.StoredFile"
                },
                "job_id": {
                    "type": "string"
                }
            }
        }
    }
}